Create explore route from server
