+++ routes/explore.js
@@ -1,0 +1,52 @@
+const express = require("express");
+const router = express.Router();
+
+router.get("/explore", (req, res) => {
+  res.render("explore", {
+    title: "Nature Nexus - Explore",
+    error: null,
+    currentPage: "explore",
+    user: req.session.user || null,
+  });
+});
+
+router.get("/openweathermap/:lat/:lon", async (req, res) => {
+  try {
+    const { lat, lon } = req.params;
+    const openweathermapAPI = process.env.OPENWEATHERMAP_API_KEY;
+    const openweathermapUrl = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=${process.env.OPENWEATHERMAP_API_KEY}`;
+    const weatherRes = await fetch(openweathermapUrl);
+    const weatherData = await weatherRes.json();
+    res.json(weatherData);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch weather data" });
+  }
+});
+
+router.get("/tiles/:z/:x/:y", async (req, res) => {
+  const { z, x, y } = req.params;
+  const tileUrl = `https://tile.jawg.io/jawg-streets/${z}/${x}/${y}.png?access-token=${process.env.JAWG_API}`;
+  try {
+    const tileRes = await fetch(tileUrl);
+
+    if (!tileRes.ok) {
+      console.error(
+        `Tile fetch failed (${tileRes.status}): ${tileRes.statusText}`
+      );
+      const errorText = await tileRes.text();
+      console.error("Tile error response:", errorText);
+      return res.status(500).send("Tile service returned an error.");
+    }
+
+    const contentType = tileRes.headers.get("content-type");
+    const buffer = await tileRes.arrayBuffer();
+
+    res.set("Content-Type", contentType || "image/png");
+    res.send(Buffer.from(buffer));
+  } catch (err) {
+    console.error("Tile fetch threw error:", err);
+    res.status(500).send("Internal fetch error.");
+  }
+});
+
+module.exports = router;
