+++ server.js
@@ -41,6 +41,7 @@ const identifyRoutes = require("./routes/identify");
 const articlesRoutes = require("./routes/articles");
 const collectionsRoutes = require("./routes/collections");
 const settingsRoutes = require("./routes/settings");
+const exploreRoutes = require("./routes/explore")
 
 // Use routes
 app.use(authRoutes);
@@ -48,6 +49,7 @@ app.use(identifyRoutes);
 app.use(articlesRoutes);
 app.use(collectionsRoutes);
 app.use(settingsRoutes);
+app.use(exploreRoutes);
 
 app.get("/", (req, res) => {
   res.render("index", {
@@ -63,54 +65,6 @@ app.get("/index", (req, res) => {
     user: req.session.user || null,
     currentPage: "home",
   });
-});
-
-app.get("/explore", (req, res) => {
-  res.render("explore", {
-    title: "Nature Nexus - Explore",
-    error: null,
-    currentPage: "explore",
-    user: req.session.user || null,
-  });
-});
-
-app.get("/openweathermap/:lat/:lon", async (req, res) => {
-  try {
-    const { lat, lon } = req.params;
-    const openweathermapAPI = process.env.OPENWEATHERMAP_API_KEY;
-    const openweathermapUrl = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=${process.env.OPENWEATHERMAP_API_KEY}`;
-    const weatherRes = await fetch(openweathermapUrl);
-    const weatherData = await weatherRes.json();
-    res.json(weatherData);
-  } catch (error) {
-    res.status(500).json({ error: "Failed to fetch weather data" });
-  }
-});
-
-app.get("/tiles/:z/:x/:y", async (req, res) => {
-  const { z, x, y } = req.params;
-  const tileUrl = `https://tile.jawg.io/jawg-streets/${z}/${x}/${y}.png?access-token=${process.env.JAWG_API}`;
-  try {
-    const tileRes = await fetch(tileUrl);
-
-    if (!tileRes.ok) {
-      console.error(
-        `Tile fetch failed (${tileRes.status}): ${tileRes.statusText}`
-      );
-      const errorText = await tileRes.text();
-      console.error("Tile error response:", errorText);
-      return res.status(500).send("Tile service returned an error.");
-    }
-
-    const contentType = tileRes.headers.get("content-type");
-    const buffer = await tileRes.arrayBuffer();
-
-    res.set("Content-Type", contentType || "image/png");
-    res.send(Buffer.from(buffer));
-  } catch (err) {
-    console.error("Tile fetch threw error:", err);
-    res.status(500).send("Internal fetch error.");
-  }
 });
 
 // Build filter helper function
